#88 package refactoring and implement style attributes
