+++ api/common/src/main/scala/de/htwg/zeta/common/model/MetaModel.scala
@@ -1,0 +1,12 @@
+package de.htwg.zeta.common.model
+
+import de.htwg.zeta.common.model.diagram.Diagram
+import de.htwg.zeta.common.model.shape.Shape
+import de.htwg.zeta.common.model.style.Style
+
+case class MetaModel(
+    id: String,
+    diagrams: List[Diagram],
+    styles: List[Style],
+    shapes: List[Shape]
+)
