+++ api/common/src/main/scala/de/htwg/zeta/common/model/diagram/Diagram.scala
@@ -1,0 +1,3 @@
+package de.htwg.zeta.common.model.diagram
+
+class Diagram()
