+++ api/common/src/main/scala/de/htwg/zeta/common/model/shape/Shape.scala
@@ -1,0 +1,3 @@
+package de.htwg.zeta.common.model.shape
+
+class Shape ()
