+++ api/common/src/main/scala/de/htwg/zeta/common/model/style/Background.scala
@@ -1,0 +1,3 @@
+package de.htwg.zeta.common.model.style
+
+case class Background(color: Color)
