+++ api/common/src/main/scala/de/htwg/zeta/common/model/style/Color.scala
@@ -1,0 +1,3 @@
+package de.htwg.zeta.common.model.style
+
+class Color(string: String)
