+++ api/common/src/main/scala/de/htwg/zeta/common/model/style/Font.scala
@@ -1,0 +1,10 @@
+package de.htwg.zeta.common.model.style
+
+case class Font(
+    bold: Boolean,
+    color: Color,
+    italic: Boolean,
+    name: String,
+    size: Int,
+    transparent: Boolean
+)
