+++ api/common/src/main/scala/de/htwg/zeta/common/model/style/Line.scala
@@ -1,0 +1,16 @@
+package de.htwg.zeta.common.model.style
+
+case class Line(
+    color: Color,
+    style: LineStyle,
+    transparent: Boolean,
+    width: Int
+)
+
+sealed trait LineStyle {
+  def value: String
+}
+
+case class Dashed(value: String = "DASH") extends LineStyle
+
+// TODO implement more style types
