+++ api/common/src/main/scala/de/htwg/zeta/common/model/style/Style.scala
@@ -1,0 +1,8 @@
+package de.htwg.zeta.common.model.style
+
+case class Style(
+    name: String,
+    background: Background,
+    font: Font,
+    transparency: Double
+)
