--- api/common/src/main/scala/de/htwg/zeta/common/models/Diagram.scala
@@ -1,3 +1,0 @@
-package de.htwg.zeta.common.models
-
-class Diagram()
