--- api/common/src/main/scala/de/htwg/zeta/common/models/MetaModel.scala
@@ -1,8 +1,0 @@
-package de.htwg.zeta.common.models
-
-case class MetaModel(
-    id: String,
-    diagrams: List[Diagram],
-    styles: List[Style],
-    shapes: List[Shape]
-)
