--- api/common/src/main/scala/de/htwg/zeta/common/models/Shape.scala
@@ -1,3 +1,0 @@
-package de.htwg.zeta.common.models
-
-class Shape ()
